+++ HabboHotel/Rooms/IRoomItemPersistenceService.cs
@@ -1,0 +1,8 @@
+using Plus.HabboHotel.Items;
+
+namespace Plus.HabboHotel.Rooms;
+
+public interface IRoomItemPersistenceService
+{
+    void SaveMovedItems(IEnumerable<Item> items);
+}